Network.HTTP2.Priority.Internal

Documentation

type Key = Int

type Weight = Int

type Deficit = Word

data Precedence

Internal representation of priority in priority queues. The precedence of a dequeued entry should be specified to enqueue when the entry is enqueued again.

Constructors

Precedence

Fields

deficit :: Deficit

weight :: Weight

dependency :: Key

newPrecedence :: Weight -> Precedence

For test only

data PriorityQueue a

Constructors

PriorityQueue

Fields

baseDeficit :: Deficit

queue :: Heap a

type Heap a = IntPSQ Precedence a

empty :: PriorityQueue a

isEmpty :: PriorityQueue a -> Bool

enqueue :: Key -> Precedence -> a -> PriorityQueue a -> PriorityQueue a

dequeue :: PriorityQueue a -> Maybe (Key, Precedence, a, PriorityQueue a)

delete :: Key -> PriorityQueue a -> (Maybe a, PriorityQueue a)