Network.HPACK

Description

HPACK(https://tools.ietf.org/html/rfc7541) encoding and decoding a header list.

Encoding and decoding

encodeHeader

Arguments

:: EncodeStrategy

-> Size
The size of a temporary buffer.

-> DynamicTable

-> HeaderList

-> IO ByteString
An HPACK format

Converting HeaderList to the HPACK format. This function has overhead of allocating/freeing a temporary buffer. BufferOverrun will be thrown if the temporary buffer is too small.

decodeHeader

Arguments

:: DynamicTable

-> ByteString
An HPACK format

-> IO HeaderList

Converting the HPACK format to HeaderList.

Headers are decoded as is.

DecodeError would be thrown if the HPACK format is broken.

BufferOverrun will be thrown if the temporary buffer for Huffman decoding is too small.

Encoding and decoding with token

encodeTokenHeader

Arguments

:: Buffer

-> BufferSize

-> EncodeStrategy

-> Bool
True at the first time, False when continued.

-> DynamicTable

-> TokenHeaderList

-> IO (TokenHeaderList, Int)
Leftover, filled length

Converting TokenHeaderList to the HPACK format directly in the buffer.

4th argument is relating to dynamic table size update. When calling this function for a new TokenHeaderList, it must be True. If True and set by setLimitForEncoding, dynamic table size update is generated at the beginning of the HPACK format.

The return value is a pair of leftover TokenHeaderList and how many bytes are filled in the buffer. If the leftover is empty, the encoding is finished. Otherwise, this function should be called with it again. 4th argument must be False.

decodeTokenHeader

Arguments

:: DynamicTable

-> ByteString
An HPACK format

-> IO (TokenHeaderList, ValueTable)

Converting the HPACK format to TokenHeaderList and ValueTable.

Multiple values of Cookie: are concatenated.

If a pseudo header appears multiple times, IllegalHeaderName is thrown.

If unknown pseudo headers appear, IllegalHeaderName is thrown.

If pseudo headers are found after normal headers, IllegalHeaderName is thrown.

If a header key contains capital letters, IllegalHeaderName is thrown.

DecodeError would be thrown if the HPACK format is broken.

BufferOverrun will be thrown if the temporary buffer for Huffman decoding is too small.

DynamicTable

data DynamicTable

Type for dynamic table.

defaultDynamicTableSize :: Int

Default dynamic table size. The value is 4,096 bytes: an array has 128 entries.

>>> defaultDynamicTableSize
4096

newDynamicTableForEncoding

Arguments

:: Size
The dynamic table size

-> IO DynamicTable

Creating DynamicTable for encoding.

newDynamicTableForDecoding

Arguments

:: Size
The dynamic table size

-> Size
The size of temporary buffer for Huffman decoding

-> IO DynamicTable

Creating DynamicTable for decoding.

clearDynamicTable :: DynamicTable -> IO ()

Clearing DynamicTable. Currently, this frees the temporary buffer for Huffman decoding.

withDynamicTableForEncoding

Arguments

:: Size
The dynamic table size

-> (DynamicTable -> IO a)

-> IO a

Creating DynamicTable for encoding, performing the action and clearing the DynamicTable.

withDynamicTableForDecoding

Arguments

:: Size
The dynamic table size

-> Size
The size of temporary buffer for Huffman

-> (DynamicTable -> IO a)

-> IO a

Creating DynamicTable for decoding, performing the action and clearing the DynamicTable.

setLimitForEncoding :: Size -> DynamicTable -> IO ()

When SETTINGS_HEADER_TABLE_SIZE is received from a peer, its value should be set by this function.

Strategy for encoding

data CompressionAlgo

Compression algorithms for HPACK encoding.

Constructors

Naive
No compression

Static
Using indices in the static table only

Linear
Using indices

data EncodeStrategy

Strategy for HPACK encoding.

Constructors

EncodeStrategy

Fields

compressionAlgo :: !CompressionAlgo

Which compression algorithm is used.

useHuffman :: !Bool

Whether or not to use Huffman encoding for strings.

defaultEncodeStrategy :: EncodeStrategy

Default EncodeStrategy.

>>> defaultEncodeStrategy
EncodeStrategy {compressionAlgo = Linear, useHuffman = False}

Errors

data DecodeError

Errors for decoder.

Constructors

IndexOverrun Index
Index is out of range

EosInTheMiddle
Eos appears in the middle of huffman string

IllegalEos
Non-eos appears in the end of huffman string

TooLongEos
Eos of huffman string is more than 7 bits

EmptyEncodedString
Encoded string has no length

TooSmallTableSize
A peer set the dynamic table size less than 32

TooLargeTableSize
A peer tried to change the dynamic table size over the limit

IllegalTableSizeUpdate
Table size update at the non-beginning

HeaderBlockTruncated

IllegalHeaderName

data BufferOverrun

Constructors

BufferOverrun
The buffer size is not enough

Headers

type HeaderList = [Header]

Header list.

type Header = (HeaderName, HeaderValue)

Header.

type HeaderName = ByteString

Header name.

type HeaderValue = ByteString

Header value.

type TokenHeaderList = [TokenHeader]

type TokenHeader = (Token, HeaderValue)

Value table

type ValueTable = Array Int (Maybe HeaderValue)

An array for HeaderValue.

getHeaderValue :: Token -> ValueTable -> Maybe HeaderValue

Accessing HeaderValue with Token.

toHeaderTable :: [(CI HeaderName, HeaderValue)] -> IO (TokenHeaderList, ValueTable)

Converting a header list of the http-types style to TokenHeaderList and ValueTable.

Basic types

type Size = Int

Size in bytes.

type Index = Int

Index for table.

type Buffer = Ptr Word8

Buffer type.

type BufferSize = Int

The size of buffer.