Network.HTTP2

Description

Framing in HTTP/2.

Frame

data Frame

The data type for HTTP/2 frames.

Constructors

Frame

Fields

frameHeader :: FrameHeader

framePayload :: FramePayload

data FrameHeader

The data type for HTTP/2 frame headers.

Constructors

FrameHeader

Fields

payloadLength :: Int

flags :: FrameFlags

streamId :: StreamId

data FramePayload

The data type for HTTP/2 frame payloads.

Constructors

DataFrame ByteString

HeadersFrame (Maybe Priority) HeaderBlockFragment

PriorityFrame Priority

RSTStreamFrame ErrorCodeId

SettingsFrame SettingsList

PushPromiseFrame StreamId HeaderBlockFragment

PingFrame ByteString

GoAwayFrame StreamId ErrorCodeId ByteString

WindowUpdateFrame WindowSize

ContinuationFrame HeaderBlockFragment

UnknownFrame FrameType ByteString

isPaddingDefined :: FramePayload -> Bool

Checking if padding is defined in this frame type.

Encoding

encodeFrame :: EncodeInfo -> FramePayload -> ByteString

Encoding an HTTP/2 frame to ByteString. This function is not efficient enough for high performace program because of the concatenation of ByteString.

>>> encodeFrame (encodeInfo id 1) (DataFrame "body")
"\NUL\NUL\EOT\NUL\NUL\NUL\NUL\NUL\SOHbody"

encodeFrameChunks :: EncodeInfo -> FramePayload -> [ByteString]

Encoding an HTTP/2 frame to [ByteString]. This is suitable for sendMany.

encodeFrameHeader :: FrameTypeId -> FrameHeader -> ByteString

Encoding an HTTP/2 frame header. The frame header must be completed.

encodeFrameHeaderBuf :: FrameTypeId -> FrameHeader -> Ptr Word8 -> IO ()

Writing an encoded HTTP/2 frame header to the buffer. The length of the buffer must be larger than or equal to 9 bytes.

encodeFramePayload :: EncodeInfo -> FramePayload -> (FrameHeader, [ByteString])

Encoding an HTTP/2 frame payload. This returns a complete frame header and chunks of payload.

data EncodeInfo

Auxiliary information for frame encoding.

Constructors

EncodeInfo

Fields

encodeFlags :: FrameFlags

Flags to be set in a frame header

encodeStreamId :: StreamId

Stream id to be set in a frame header

encodePadding :: Maybe Padding

Padding if any. In the case where this value is set but the priority flag is not set, this value gets preference over the priority flag. So, if this value is set, the priority flag is also set.

encodeInfo

Arguments

:: (FrameFlags -> FrameFlags)

-> Int
stream identifier

-> EncodeInfo

A smart builder of EncodeInfo.

>>> encodeInfo setAck 0
EncodeInfo {encodeFlags = 1, encodeStreamId = 0, encodePadding = Nothing}

Decoding

decodeFrame

Arguments

:: Settings
HTTP/2 settings

-> ByteString
Input byte-stream

-> Either HTTP2Error Frame
Decoded frame

Decoding an HTTP/2 frame to ByteString. The second argument must be include the entire of frame. So, this function is not useful for real applications but useful for testing.

decodeFrameHeader :: ByteString -> (FrameTypeId, FrameHeader)

Decoding an HTTP/2 frame header. Must supply 9 bytes.

checkFrameHeader :: Settings -> (FrameTypeId, FrameHeader) -> Either HTTP2Error (FrameTypeId, FrameHeader)

Checking a frame header and reporting an error if any.

>>> checkFrameHeader defaultSettings (FrameData,(FrameHeader 100 0 0))
Left (ConnectionError ProtocolError "cannot used in control stream")

Decoding payload

decodeFramePayload :: FrameTypeId -> FramePayloadDecoder

Decoding an HTTP/2 frame payload.

type FramePayloadDecoder = FrameHeader -> ByteString -> Either HTTP2Error FramePayload

decodeDataFrame :: FramePayloadDecoder

decodeHeadersFrame :: FramePayloadDecoder

decodePriorityFrame :: FramePayloadDecoder

decoderstStreamFrame :: FramePayloadDecoder

decodeSettingsFrame :: FramePayloadDecoder

decodePushPromiseFrame :: FramePayloadDecoder

decodePingFrame :: FramePayloadDecoder

decodeGoAwayFrame :: FramePayloadDecoder

decodeWindowUpdateFrame :: FramePayloadDecoder

decodeContinuationFrame :: FramePayloadDecoder

Frame type ID

data FrameTypeId

Constructors

FrameData

FrameHeaders

FramePriority

FrameRSTStream

FrameSettings

FramePushPromise

FramePing

FrameGoAway

FrameWindowUpdate

FrameContinuation

FrameUnknown FrameType

framePayloadToFrameTypeId :: FramePayload -> FrameTypeId

Getting FrameType from FramePayload.

>>> framePayloadToFrameTypeId (DataFrame "body")
FrameData

Frame type

type FrameType = Word8

fromFrameTypeId :: FrameTypeId -> FrameType

>>> fromFrameTypeId FrameData
0
>>> fromFrameTypeId FrameContinuation
9
>>> fromFrameTypeId (FrameUnknown 10)
10

toFrameTypeId :: FrameType -> FrameTypeId

>>> toFrameTypeId 0
FrameData
>>> toFrameTypeId 9
FrameContinuation
>>> toFrameTypeId 10
FrameUnknown 10

Types

type HeaderBlockFragment = ByteString

type Padding = ByteString

type Weight = Int

data Priority

Constructors

Priority

Fields

exclusive :: Bool

streamDependency :: StreamId

weight :: Weight

defaultPriority :: Priority

Default priority which depends on stream 0.

>>> defaultPriority
Priority {exclusive = False, streamDependency = 0, weight = 16}

highestPriority :: Priority

Highest priority which depends on stream 0.

>>> highestPriority
Priority {exclusive = False, streamDependency = 0, weight = 256}

Stream identifier

type StreamId = Int

isControl :: StreamId -> Bool

>>> isControl 0
True
>>> isControl 1
False

isRequest :: StreamId -> Bool

>>> isRequest 0
False
>>> isRequest 1
True

isResponse :: StreamId -> Bool

>>> isResponse 0
False
>>> isResponse 2
True

Stream identifier related

testExclusive :: Int -> Bool

setExclusive :: Int -> Int

clearExclusive :: Int -> Int

Flags

type FrameFlags = Word8

defaultFlags :: FrameFlags

>>> defaultFlags
0

testEndStream :: FrameFlags -> Bool

>>> testEndStream 0x1
True

testAck :: FrameFlags -> Bool

>>> testAck 0x1
True

testEndHeader :: FrameFlags -> Bool

>>> testEndHeader 0x4
True

testPadded :: FrameFlags -> Bool

>>> testPadded 0x8
True

testPriority :: FrameFlags -> Bool

>>> testPriority 0x20
True

setEndStream :: FrameFlags -> FrameFlags

>>> setEndStream 0
1

setAck :: FrameFlags -> FrameFlags

>>> setAck 0
1

setEndHeader :: FrameFlags -> FrameFlags

>>> setEndHeader 0
4

setPadded :: FrameFlags -> FrameFlags

>>> setPadded 0
8

setPriority :: FrameFlags -> FrameFlags

>>> setPriority 0
32

SettingsList

type SettingsList = [(SettingsKeyId, SettingsValue)]

Settings containing raw values.

type SettingsValue = Int

data SettingsKeyId

Constructors

SettingsHeaderTableSize

SettingsEnablePush

SettingsMaxConcurrentStreams

SettingsInitialWindowSize

SettingsMaxFrameSize

SettingsMaxHeaderBlockSize

fromSettingsKeyId :: SettingsKeyId -> Word16

>>> fromSettingsKeyId SettingsHeaderTableSize
1
>>> fromSettingsKeyId SettingsMaxHeaderBlockSize
6

toSettingsKeyId :: Word16 -> Maybe SettingsKeyId

>>> toSettingsKeyId 0
Nothing
>>> toSettingsKeyId 1
Just SettingsHeaderTableSize
>>> toSettingsKeyId 6
Just SettingsMaxHeaderBlockSize
>>> toSettingsKeyId 7
Nothing

checkSettingsList :: SettingsList -> Maybe HTTP2Error

Checking SettingsList and reporting an error if any.

>>> checkSettingsList [(SettingsEnablePush,2)]
Just (ConnectionError ProtocolError "enable push must be 0 or 1")

Settings

data Settings

Cooked version of settings. This is suitable to be stored in a HTTP/2 context.

Constructors

Settings

Fields

headerTableSize :: Int

enablePush :: Bool

maxConcurrentStreams :: Maybe Int

initialWindowSize :: WindowSize

maxFrameSize :: Int

maxHeaderBlockSize :: Maybe Int

defaultSettings :: Settings

The default settings.

>>> defaultSettings
Settings {headerTableSize = 4096, enablePush = True, maxConcurrentStreams = Nothing, initialWindowSize = 65535, maxFrameSize = 16384, maxHeaderBlockSize = Nothing}

updateSettings :: Settings -> SettingsList -> Settings

Updating settings.

>>> updateSettings defaultSettings [(SettingsEnablePush,0),(SettingsMaxHeaderBlockSize,200)]
Settings {headerTableSize = 4096, enablePush = False, maxConcurrentStreams = Nothing, initialWindowSize = 65535, maxFrameSize = 16384, maxHeaderBlockSize = Just 200}

Window

type WindowSize = Int

defaultInitialWindowSize :: WindowSize

The default initial window size.

>>> defaultInitialWindowSize
65535

maxWindowSize :: WindowSize

The maximum window size.

>>> maxWindowSize
2147483647

isWindowOverflow :: WindowSize -> Bool

Checking if a window size exceeds the maximum window size.

>>> isWindowOverflow 10
False
>>> isWindowOverflow maxWindowSize
False
>>> isWindowOverflow (maxWindowSize + 1)
True

Misc

recommendedConcurrency :: Int

Default concurrency.

>>> recommendedConcurrency
100

Error code

type ErrorCode = Word32

data ErrorCodeId

Constructors

NoError

ProtocolError

InternalError

FlowControlError

SettingsTimeout

StreamClosed

FrameSizeError

RefusedStream

Cancel

CompressionError

ConnectError

EnhanceYourCalm

InadequateSecurity

HTTP11Required

UnknownErrorCode ErrorCode

fromErrorCodeId :: ErrorCodeId -> ErrorCode

>>> fromErrorCodeId NoError
0
>>> fromErrorCodeId InadequateSecurity
12

toErrorCodeId :: ErrorCode -> ErrorCodeId

>>> toErrorCodeId 0
NoError
>>> toErrorCodeId 0xc
InadequateSecurity
>>> toErrorCodeId 0xe
UnknownErrorCode 14

Error

data HTTP2Error

The connection error or the stream error.

Constructors

ConnectionError ErrorCodeId ByteString

StreamError ErrorCodeId StreamId

errorCodeId :: HTTP2Error -> ErrorCodeId

Obtaining ErrorCodeId from HTTP2Error.

Magic

connectionPreface :: ByteString

The preface of HTTP/2.

>>> connectionPreface
"PRI * HTTP/2.0\r\n\r\nSM\r\n\r\n"

connectionPrefaceLength :: Int

Length of the preface.

>>> connectionPrefaceLength
24

frameHeaderLength :: Int

The length of HTTP/2 frame header.

>>> frameHeaderLength
9

maxPayloadLength :: Int

The maximum length of HTTP/2 payload.

>>> maxPayloadLength
16384