Network.HTTP2.Client

Description

HTTP/2 client library.

Example:

{-# LANGUAGE OverloadedStrings #-}
{-# LANGUAGE RankNTypes #-}

module Main where

import Control.Concurrent.Async
import qualified Control.Exception as E
import qualified Data.ByteString.Char8 as C8
import Network.HTTP.Types
import Network.Run.TCP (runTCPClient) -- network-run

import Network.HTTP2.Client

serverName :: String
serverName = "127.0.0.1"

main :: IO ()
main = runTCPClient serverName "80" $ runHTTP2Client serverName
  where
    cliconf host = defaultClientConfig { authority = C8.pack host }
    runHTTP2Client host s = E.bracket (allocSimpleConfig s 4096)
                                      freeSimpleConfig
                                      (\conf -> run (cliconf host) conf client)
    client :: Client ()
    client sendRequest _aux = do
        let req0 = requestNoBody methodGet "/" []
            client0 = sendRequest req0 $ \rsp -> do
                print rsp
                getResponseBodyChunk rsp >>= C8.putStrLn
            req1 = requestNoBody methodGet "/foo" []
            client1 = sendRequest req1 $ \rsp -> do
                print rsp
                getResponseBodyChunk rsp >>= C8.putStrLn
        ex <- E.try $ concurrently_ client0 client1
        case ex of
          Left  e  -> print (e :: HTTP2Error)
          Right () -> putStrLn "OK"

Runner

run :: ClientConfig -> Config -> Client a -> IO a

Running HTTP/2 client.

Client configuration

type Scheme = ByteString

"http" or "https".

type Authority = ByteString

Authority.

data ClientConfig

Client configuration

defaultClientConfig :: ClientConfig

The default client config.

The authority field will be used to set the HTTP2 :authority pseudo-header. In most cases you will want to override it to be equal to host.

Further background on authority: RFC 3986 also allows host:port, and most servers will accept this too. However, when using TLS, many servers will expect the TLS SNI server name and the :authority pseudo-header to be equal, and for TLS SNI the server name should not include the port. Note that HTTP2 explicitly disallows using userinfo@ as part of the authority.

>>> defaultClientConfig
ClientConfig {scheme = "http", authority = "localhost", cacheLimit = 64, connectionWindowSize = 1048576, settings = Settings {headerTableSize = 4096, enablePush = True, maxConcurrentStreams = Just 64, initialWindowSize = 262144, maxFrameSize = 16384, maxHeaderListSize = Nothing}}

scheme :: ClientConfig -> Scheme

https or http

authority :: ClientConfig -> Authority

Server name

cacheLimit :: ClientConfig -> Int

The maximum number of incoming streams on the net

connectionWindowSize :: ClientConfig -> WindowSize

The window size of connection.

settings :: ClientConfig -> Settings

Settings

HTTP/2 setting

data Settings

HTTP/2 settings. See https://datatracker.ietf.org/doc/html/rfc9113#name-defined-settings.

defaultSettings :: Settings

The default settings.

>>> defaultSettings
Settings {headerTableSize = 4096, enablePush = True, maxConcurrentStreams = Just 64, initialWindowSize = 262144, maxFrameSize = 16384, maxHeaderListSize = Nothing}

headerTableSize :: Settings -> Int

SETTINGS_HEADER_TABLE_SIZE

enablePush :: Settings -> Bool

SETTINGS_ENABLE_PUSH

maxConcurrentStreams :: Settings -> Maybe Int

SETTINGS_MAX_CONCURRENT_STREAMS

initialWindowSize :: Settings -> WindowSize

SETTINGS_INITIAL_WINDOW_SIZE

maxFrameSize :: Settings -> Int

SETTINGS_MAX_FRAME_SIZE

maxHeaderListSize :: Settings -> Maybe Int

SETTINGS_MAX_HEADER_LIST_SIZE

Common configuration

data Config

HTTP/2 configuration.

Constructors

Config

Fields

confWriteBuffer :: Buffer

This is used only by frameSender. This MUST be freed after frameSender is terminated.

confBufferSize :: BufferSize

The size of the write buffer. We assume that the read buffer is the same size. So, this value is announced via SETTINGS_MAX_FRAME_SIZE to the peer.

confSendAll :: ByteString -> IO ()

confReadN :: Int -> IO ByteString

confPositionReadMaker :: PositionReadMaker

confTimeoutManager :: Manager

confMySockAddr :: SockAddr

This is copied into Aux, if exist, on server.

confPeerSockAddr :: SockAddr

This is copied into Aux, if exist, on server.

allocSimpleConfig :: Socket -> BufferSize -> IO Config

Making simple configuration whose IO is not efficient. A write buffer is allocated internally.

freeSimpleConfig :: Config -> IO ()

Deallocating the resource of the simple configuration.

HTTP/2 client

type Client a = (forall b. Request -> (Response -> IO b) -> IO b) -> Aux -> IO a

Client type.

Request

data Request

Request from client.

Creating request

requestNoBody :: Method -> Path -> RequestHeaders -> Request

Creating request without body.

requestFile :: Method -> Path -> RequestHeaders -> FileSpec -> Request

Creating request with file.

requestStreaming :: Method -> Path -> RequestHeaders -> ((Builder -> IO ()) -> IO () -> IO ()) -> Request

Creating request with streaming.

requestStreamingUnmask :: Method -> Path -> RequestHeaders -> ((forall x. IO x -> IO x) -> (Builder -> IO ()) -> IO () -> IO ()) -> Request

Like requestStreaming, but run the action with exceptions masked

requestBuilder :: Method -> Path -> RequestHeaders -> Builder -> Request

Creating request with builder.

Trailers maker

type TrailersMaker = Maybe ByteString -> IO NextTrailersMaker

Trailers maker. A chunks of the response body is passed with Just. The maker should update internal state with the ByteString and return the next trailers maker. When response body reaches its end, Nothing is passed and the maker should generate trailers. An example:

{-# LANGUAGE BangPatterns #-}
import Data.ByteString (ByteString)
import qualified Data.ByteString.Char8 as C8
import Crypto.Hash (Context, SHA1) -- cryptonite
import qualified Crypto.Hash as CH

-- Strictness is important for Context.
trailersMaker :: Context SHA1 -> Maybe ByteString -> IO NextTrailersMaker
trailersMaker ctx Nothing = return $ Trailers [("X-SHA1", sha1)]
  where
    !sha1 = C8.pack $ show $ CH.hashFinalize ctx
trailersMaker ctx (Just bs) = return $ NextTrailersMaker $ trailersMaker ctx'
  where
    !ctx' = CH.hashUpdate ctx bs

Usage example:

let h2rsp = responseFile ...
    maker = trailersMaker (CH.hashInit :: Context SHA1)
    h2rsp' = setResponseTrailersMaker h2rsp maker

data NextTrailersMaker

Either the next trailers maker or final trailers.

Constructors

NextTrailersMaker TrailersMaker

Trailers [Header]

defaultTrailersMaker :: TrailersMaker

TrailersMake to create no trailers.

setRequestTrailersMaker :: Request -> TrailersMaker -> Request

Setting TrailersMaker to Response.

Response

data Response

Response from server.

Accessing response

responseStatus :: Response -> Maybe Status

Getting the status of a response.

responseHeaders :: Response -> HeaderTable

Getting the headers from a response.

responseBodySize :: Response -> Maybe Int

Getting the body size from a response.

getResponseBodyChunk :: Response -> IO ByteString

Reading a chunk of the response body. An empty ByteString returned when finished.

getResponseTrailers :: Response -> IO (Maybe HeaderTable)

Reading response trailers. This function must be called after getResponseBodyChunk returns an empty.

Aux

data Aux

Additional information.

auxPossibleClientStreams :: Aux -> IO Int

How many streams can be created without blocking.

Types

type Method = ByteString

HTTP method (flat ByteString type).

type Path = ByteString

Path.

data FileSpec

File specification.

Constructors

FileSpec FilePath FileOffset ByteCount

type FileOffset = Int64

Offset for file.

type ByteCount = Int64

How many bytes to read

Error

data HTTP2Error

The connection error or the stream error. Stream errors are treated as connection errors since there are no good recovery ways. ErrorCode in connection errors should be the highest stream identifier but in this implementation it identifies the stream that caused this error.

Constructors

ConnectionIsClosed

ConnectionIsTimeout

ConnectionErrorIsReceived ErrorCode StreamId ReasonPhrase

ConnectionErrorIsSent ErrorCode StreamId ReasonPhrase

StreamErrorIsReceived ErrorCode StreamId

StreamErrorIsSent ErrorCode StreamId ReasonPhrase

BadThingHappen SomeException

GoAwayIsSent

type ReasonPhrase = ShortByteString

newtype ErrorCode

The type for raw error code.

Constructors

ErrorCode Word32

Bundled Patterns

pattern NoError :: ErrorCode
The type for error code. See https://www.rfc-editor.org/rfc/rfc9113#ErrorCodes.

pattern ProtocolError :: ErrorCode

pattern InternalError :: ErrorCode

pattern FlowControlError :: ErrorCode

pattern SettingsTimeout :: ErrorCode

pattern StreamClosed :: ErrorCode

pattern FrameSizeError :: ErrorCode

pattern RefusedStream :: ErrorCode

pattern Cancel :: ErrorCode

pattern CompressionError :: ErrorCode

pattern ConnectError :: ErrorCode

pattern EnhanceYourCalm :: ErrorCode

pattern InadequateSecurity :: ErrorCode

pattern HTTP11Required :: ErrorCode

RecvN

defaultReadN :: Socket -> IORef (Maybe ByteString) -> Int -> IO ByteString

Naive implementation for readN.

Position read for files

type PositionReadMaker = FilePath -> IO (PositionRead, Sentinel)

Making a position read and its closer.

type PositionRead = FileOffset -> ByteCount -> Buffer -> IO ByteCount

Position read for files.

data Sentinel

Manipulating a file resource.

Constructors

Closer (IO ())
Closing a file resource. Its refresher is automatiaclly generated by the internal timer.

Refresher (IO ())
Refreshing a file resource while reading. Closing the file must be done by its own timer or something.

defaultPositionReadMaker :: PositionReadMaker

Position read based on Handle.