Network.HTTP2.Client

Description

HTTP/2 client library.

Example:

{-# LANGUAGE OverloadedStrings #-}
module Main where

import qualified Control.Exception as E
import Control.Concurrent (forkIO, threadDelay)
import qualified Data.ByteString.Char8 as C8
import Network.HTTP.Types
import Network.Run.TCP (runTCPClient) -- network-run

import Network.HTTP2.Client

serverName :: String
serverName = "127.0.0.1"

main :: IO ()
main = runTCPClient serverName "80" runHTTP2Client
  where
    cliconf = ClientConfig "http" (C8.pack serverName) 20
    runHTTP2Client s = E.bracket (allocSimpleConfig s 4096)
                                 freeSimpleConfig
                                 (\conf -> run cliconf conf client)
    client sendRequest = do
        let req = requestNoBody methodGet "/" []
        _ <- forkIO $ sendRequest req $ \rsp -> do
            print rsp
            getResponseBodyChunk rsp >>= C8.putStrLn
        sendRequest req $ \rsp -> do
            threadDelay 100000
            print rsp
            getResponseBodyChunk rsp >>= C8.putStrLn

Runner

run :: ClientConfig -> Config -> Client a -> IO a

Running HTTP/2 client.

type Scheme = ByteString

"http" or "https".

type Authority = ByteString

For so-called "Host:" header.

Runner arguments

data ClientConfig

Client configuration

Constructors

ClientConfig

Fields

scheme :: Scheme

https or http

authority :: Authority

Server name

cacheLimit :: Int

How many pushed responses are contained in the cache

data Config

HTTP/2 configuration.

Constructors

Config

Fields

confWriteBuffer :: Buffer

confBufferSize :: BufferSize

confSendAll :: ByteString -> IO ()

confReadN :: Int -> IO ByteString

confPositionReadMaker :: PositionReadMaker

confTimeoutManager :: Manager

allocSimpleConfig :: Socket -> BufferSize -> IO Config

Making simple configuration whose IO is not efficient. A write buffer is allocated internally.

freeSimpleConfig :: Config -> IO ()

Deallocating the resource of the simple configuration.

HTTP/2 client

type Client a = (Request -> (Response -> IO a) -> IO a) -> IO a

Client type.

Request

data Request

Request from client.

Creating request

requestNoBody :: Method -> Path -> RequestHeaders -> Request

Creating request without body.

requestFile :: Method -> Path -> RequestHeaders -> FileSpec -> Request

Creating request with file.

requestStreaming :: Method -> Path -> RequestHeaders -> ((Builder -> IO ()) -> IO () -> IO ()) -> Request

Creating request with streaming.

requestBuilder :: Method -> Path -> RequestHeaders -> Builder -> Request

Creating request with builder.

Trailers maker

type TrailersMaker = Maybe ByteString -> IO NextTrailersMaker

Trailers maker. A chunks of the response body is passed with Just. The maker should update internal state with the ByteString and return the next trailers maker. When response body reaches its end, Nothing is passed and the maker should generate trailers. An example:

{-# LANGUAGE BangPatterns #-}
import Data.ByteString (ByteString)
import qualified Data.ByteString.Char8 as C8
import Crypto.Hash (Context, SHA1) -- cryptonite
import qualified Crypto.Hash as CH

-- Strictness is important for Context.
trailersMaker :: Context SHA1 -> Maybe ByteString -> IO NextTrailersMaker
trailersMaker ctx Nothing = return $ Trailers [("X-SHA1", sha1)]
  where
    !sha1 = C8.pack $ show $ CH.hashFinalize ctx
trailersMaker ctx (Just bs) = return $ NextTrailersMaker $ trailersMaker ctx'
  where
    !ctx' = CH.hashUpdate ctx bs

Usage example:

let h2rsp = responseFile ...
    maker = trailersMaker (CH.hashInit :: Context SHA1)
    h2rsp' = setResponseTrailersMaker h2rsp maker

data NextTrailersMaker

Either the next trailers maker or final trailers.

Constructors

NextTrailersMaker TrailersMaker

Trailers [Header]

defaultTrailersMaker :: TrailersMaker

TrailersMake to create no trailers.

setRequestTrailersMaker :: Request -> TrailersMaker -> Request

Setting TrailersMaker to Response.

Response

data Response

Response from server.

Accessing response

responseStatus :: Response -> Maybe Status

Getting the status of a response.

responseHeaders :: Response -> HeaderTable

Getting the headers from a response.

responseBodySize :: Response -> Maybe Int

Getting the body size from a response.

getResponseBodyChunk :: Response -> IO ByteString

Reading a chunk of the response body. An empty ByteString returned when finished.

getResponseTrailers :: Response -> IO (Maybe HeaderTable)

Reading response trailers. This function must be called after getResponseBodyChunk returns an empty.

Types

type Method = ByteString

HTTP method (flat string type).

type Path = ByteString

Path.

data FileSpec

File specification.

Constructors

FileSpec FilePath FileOffset ByteCount

type FileOffset = Int64

Offset for file.

type ByteCount = Int64

How many bytes to read

RecvN

defaultReadN :: Socket -> IORef (Maybe ByteString) -> Int -> IO ByteString

Naive implementation for readN.

Position read for files

type PositionReadMaker = FilePath -> IO (PositionRead, Sentinel)

Making a position read and its closer.

type PositionRead = FileOffset -> ByteCount -> Buffer -> IO ByteCount

Position read for files.

data Sentinel

Manipulating a file resource.

Constructors

Closer (IO ())
Closing a file resource. Its refresher is automatiaclly generated by the internal timer.

Refresher (IO ())
Refreshing a file resource while reading. Closing the file must be done by its own timer or something.

defaultPositionReadMaker :: PositionReadMaker

Position read based on Handle.