Network.HTTP2.Internal

File

type FileOffset = Int64

Offset for file.

type ByteCount = Int64

How many bytes to read

type PositionRead = FileOffset -> ByteCount -> Buffer -> IO ByteCount

Position read for files.

data Sentinel

Manipulating a file resource.

Constructors

Closer (IO ())
Closing a file resource. Its refresher is automatiaclly generated by the internal timer.

Refresher (IO ())
Refreshing a file resource while reading. Closing the file must be done by its own timer or something.

type PositionReadMaker = FilePath -> IO (PositionRead, Sentinel)

Making a position read and its closer.

defaultPositionReadMaker :: PositionReadMaker

Position read based on Handle.

Types

type Scheme = ByteString

"http" or "https".

type Authority = ByteString

For so-called "Host:" header.

type Path = ByteString

Path.

Request and response

data InpObj

Input object

Constructors

InpObj

Fields

inpObjHeaders :: HeaderTable

Accessor for headers.

inpObjBodySize :: Maybe Int

Accessor for body length specified in content-length:.

inpObjBody :: InpBody

Accessor for body.

inpObjTrailers :: IORef (Maybe HeaderTable)

Accessor for trailers.

type InpBody = IO ByteString

data OutObj

Output object

Constructors

OutObj

Fields

outObjHeaders :: [Header]

Accessor for header.

outObjBody :: OutBody

Accessor for outObj body.

outObjTrailers :: TrailersMaker

Accessor for trailers maker.

data OutBody

Constructors

OutBodyNone

OutBodyStreaming ((Builder -> IO ()) -> IO () -> IO ())
Streaming body takes a write action and a flush action.

OutBodyBuilder Builder

OutBodyFile FileSpec

data FileSpec

File specification.

Constructors

FileSpec FilePath FileOffset ByteCount

Sender

data Next

Constructors

Next BytesFilled (Maybe DynaNext)

type BytesFilled = Int

type DynaNext = Buffer -> BufferSize -> WindowSize -> IO Next

data StreamingChunk

Constructors

StreamingFinished

StreamingFlush

StreamingBuilder Builder

fillBuilderBodyGetNext :: Builder -> DynaNext

fillFileBodyGetNext :: PositionRead -> FileOffset -> ByteCount -> IO () -> DynaNext

fillStreamBodyGetNext :: IO (Maybe StreamingChunk) -> DynaNext

Trailer

type TrailersMaker = Maybe ByteString -> IO NextTrailersMaker

Trailers maker. A chunks of the response body is passed with Just. The maker should update internal state with the ByteString and return the next trailers maker. When response body reaches its end, Nothing is passed and the maker should generate trailers. An example:

{-# LANGUAGE BangPatterns #-}
import Data.ByteString (ByteString)
import qualified Data.ByteString.Char8 as C8
import Crypto.Hash (Context, SHA1) -- cryptonite
import qualified Crypto.Hash as CH

-- Strictness is important for Context.
trailersMaker :: Context SHA1 -> Maybe ByteString -> IO NextTrailersMaker
trailersMaker ctx Nothing = return $ Trailers [("X-SHA1", sha1)]
  where
    !sha1 = C8.pack $ show $ CH.hashFinalize ctx
trailersMaker ctx (Just bs) = return $ NextTrailersMaker $ trailersMaker ctx'
  where
    !ctx' = CH.hashUpdate ctx bs

Usage example:

let h2rsp = responseFile ...
    maker = trailersMaker (CH.hashInit :: Context SHA1)
    h2rsp' = setResponseTrailersMaker h2rsp maker

defaultTrailersMaker :: TrailersMaker

TrailersMake to create no trailers.

data NextTrailersMaker

Either the next trailers maker or final trailers.

Constructors

NextTrailersMaker TrailersMaker

Trailers [Header]

runTrailersMaker :: TrailersMaker -> Buffer -> Int -> IO NextTrailersMaker

Running trailers-maker.

bufferIO buf siz $ \bs -> tlrmkr (Just bs)