Network.HTTP2.Server

Description

HTTP/2 server library.

Example:

{-# LANGUAGE OverloadedStrings #-}
module Main (main) where

import qualified Control.Exception as E
import Data.ByteString.Builder (byteString)
import Network.HTTP.Types (ok200)
import Network.Run.TCP (runTCPServer) -- network-run

import Network.HTTP2.Server

main :: IO ()
main = runTCPServer Nothing "80" runHTTP2Server
  where
    runHTTP2Server s = E.bracket (allocSimpleConfig s 4096)
                                 freeSimpleConfig
                                 (\config -> run defaultServerConfig config server)
    server _req _aux sendResponse = sendResponse response []
      where
        response = responseBuilder ok200 header body
        header = [("Content-Type", "text/plain")]
        body = byteString "Hello, world!\n"

Runner

run :: ServerConfig -> Config -> Server -> IO ()

Running HTTP/2 server.

Server configuration

data ServerConfig

Server configuration

defaultServerConfig :: ServerConfig

The default server config.

>>> defaultServerConfig
ServerConfig {numberOfWorkers = 8, connectionWindowSize = 1048576, settings = Settings {headerTableSize = 4096, enablePush = True, maxConcurrentStreams = Just 64, initialWindowSize = 262144, maxFrameSize = 16384, maxHeaderListSize = Nothing}}

numberOfWorkers :: ServerConfig -> Int

The number of workers

connectionWindowSize :: ServerConfig -> WindowSize

The window size of incoming streams

settings :: ServerConfig -> Settings

Settings

HTTP/2 setting

data Settings

HTTP/2 settings. See https://datatracker.ietf.org/doc/html/rfc9113#name-defined-settings.

defaultSettings :: Settings

The default settings.

>>> defaultSettings
Settings {headerTableSize = 4096, enablePush = True, maxConcurrentStreams = Just 64, initialWindowSize = 262144, maxFrameSize = 16384, maxHeaderListSize = Nothing}

headerTableSize :: Settings -> Int

SETTINGS_HEADER_TABLE_SIZE

enablePush :: Settings -> Bool

SETTINGS_ENABLE_PUSH

maxConcurrentStreams :: Settings -> Maybe Int

SETTINGS_MAX_CONCURRENT_STREAMS

initialWindowSize :: Settings -> WindowSize

SETTINGS_INITIAL_WINDOW_SIZE

maxFrameSize :: Settings -> Int

SETTINGS_MAX_FRAME_SIZE

maxHeaderListSize :: Settings -> Maybe Int

SETTINGS_MAX_HEADER_LIST_SIZE

Common configuration

data Config

HTTP/2 configuration.

Constructors

Config

Fields

confWriteBuffer :: Buffer

This is used only by frameSender. This MUST be freed after frameSender is terminated.

confBufferSize :: BufferSize

The size of the write buffer. We assume that the read buffer is the same size. So, this value is announced via SETTINGS_MAX_FRAME_SIZE to the peer.

confSendAll :: ByteString -> IO ()

confReadN :: Int -> IO ByteString

confPositionReadMaker :: PositionReadMaker

confTimeoutManager :: Manager

confMySockAddr :: SockAddr

This is copied into Aux, if exist, on server.

confPeerSockAddr :: SockAddr

This is copied into Aux, if exist, on server.

allocSimpleConfig :: Socket -> BufferSize -> IO Config

Making simple configuration whose IO is not efficient. A write buffer is allocated internally.

freeSimpleConfig :: Config -> IO ()

Deallocating the resource of the simple configuration.

HTTP/2 server

type Server = Request -> Aux -> (Response -> [PushPromise] -> IO ()) -> IO ()

Server type. Server takes a HTTP request, should generate a HTTP response and push promises, then should give them to the sending function. The sending function would throw exceptions so that they can be logged.

Request

data Request

Request from client.

Accessing request

requestMethod :: Request -> Maybe Method

Getting the method from a request.

requestPath :: Request -> Maybe Path

Getting the path from a request.

requestAuthority :: Request -> Maybe Authority

Getting the authority from a request.

requestScheme :: Request -> Maybe Scheme

Getting the scheme from a request.

requestHeaders :: Request -> HeaderTable

Getting the headers from a request.

requestBodySize :: Request -> Maybe Int

Getting the body size from a request.

getRequestBodyChunk :: Request -> IO ByteString

Reading a chunk of the request body. An empty ByteString returned when finished.

getRequestTrailers :: Request -> IO (Maybe HeaderTable)

Reading request trailers. This function must be called after getRequestBodyChunk returns an empty.

Aux

data Aux

Additional information.

auxTimeHandle :: Aux -> Handle

Time handle for the worker processing this request and response.

auxMySockAddr :: Aux -> SockAddr

Local socket address copied from Config.

auxPeerSockAddr :: Aux -> SockAddr

Remove socket address copied from Config.

Response

data Response

Response from server.

Creating response

responseNoBody :: Status -> ResponseHeaders -> Response

Creating response without body.

responseFile :: Status -> ResponseHeaders -> FileSpec -> Response

Creating response with file.

responseStreaming :: Status -> ResponseHeaders -> ((Builder -> IO ()) -> IO () -> IO ()) -> Response

Creating response with streaming.

responseBuilder :: Status -> ResponseHeaders -> Builder -> Response

Creating response with builder.

Accessing response

responseBodySize :: Response -> Maybe Int

Getter for response body size. This value is available for file body.

Trailers maker

type TrailersMaker = Maybe ByteString -> IO NextTrailersMaker

Trailers maker. A chunks of the response body is passed with Just. The maker should update internal state with the ByteString and return the next trailers maker. When response body reaches its end, Nothing is passed and the maker should generate trailers. An example:

{-# LANGUAGE BangPatterns #-}
import Data.ByteString (ByteString)
import qualified Data.ByteString.Char8 as C8
import Crypto.Hash (Context, SHA1) -- cryptonite
import qualified Crypto.Hash as CH

-- Strictness is important for Context.
trailersMaker :: Context SHA1 -> Maybe ByteString -> IO NextTrailersMaker
trailersMaker ctx Nothing = return $ Trailers [("X-SHA1", sha1)]
  where
    !sha1 = C8.pack $ show $ CH.hashFinalize ctx
trailersMaker ctx (Just bs) = return $ NextTrailersMaker $ trailersMaker ctx'
  where
    !ctx' = CH.hashUpdate ctx bs

Usage example:

let h2rsp = responseFile ...
    maker = trailersMaker (CH.hashInit :: Context SHA1)
    h2rsp' = setResponseTrailersMaker h2rsp maker

data NextTrailersMaker

Either the next trailers maker or final trailers.

Constructors

NextTrailersMaker TrailersMaker

Trailers [Header]

defaultTrailersMaker :: TrailersMaker

TrailersMake to create no trailers.

setResponseTrailersMaker :: Response -> TrailersMaker -> Response

Setting TrailersMaker to Response.

Push promise

data PushPromise

HTTP/2 push promise or sever push. Pseudo REQUEST headers in push promise is automatically generated. Then, a server push is sent according to promiseResponse.

pushPromise :: ByteString -> Response -> Weight -> PushPromise

Creating push promise. The third argument is traditional, not used.

promiseRequestPath :: PushPromise -> ByteString

Accessor for a URL path in a push promise (a virtual request from a server). E.g. "/style/default.css".

promiseResponse :: PushPromise -> Response

Accessor for response actually pushed from a server.

Types

type Path = ByteString

Path.

type Authority = ByteString

Authority.

type Scheme = ByteString

"http" or "https".

data FileSpec

File specification.

Constructors

FileSpec FilePath FileOffset ByteCount

type FileOffset = Int64

Offset for file.

type ByteCount = Int64

How many bytes to read

RecvN

defaultReadN :: Socket -> IORef (Maybe ByteString) -> Int -> IO ByteString

Naive implementation for readN.

Position read for files

type PositionReadMaker = FilePath -> IO (PositionRead, Sentinel)

Making a position read and its closer.

type PositionRead = FileOffset -> ByteCount -> Buffer -> IO ByteCount

Position read for files.

data Sentinel

Manipulating a file resource.

Constructors

Closer (IO ())
Closing a file resource. Its refresher is automatiaclly generated by the internal timer.

Refresher (IO ())
Refreshing a file resource while reading. Closing the file must be done by its own timer or something.

defaultPositionReadMaker :: PositionReadMaker

Position read based on Handle.