Network.HTTP2.Server

Description

HTTP/2 server library.

Example:

{-# LANGUAGE OverloadedStrings #-}
module Main (main) where

import qualified Control.Exception as E
import Data.ByteString.Builder (byteString)
import Network.HTTP.Types (ok200)
import Network.Run.TCP (runTCPServer) -- network-run

import Network.HTTP2.Server

main :: IO ()
main = runTCPServer Nothing "80" $ \s _peer -> runHTTP2Server s
  where
    runHTTP2Server s = E.bracket (allocSimpleConfig s 4096)
                                 freeSimpleConfig
                                 (\config -> run config server)
    server _req _aux sendResponse = sendResponse response []
      where
        response = responseBuilder ok200 header body
        header = [("Content-Type", "text/plain")]
        body = byteString "Hello, world!\n"

Runner

run :: Config -> Server -> IO ()

Running HTTP/2 server.

Runner arguments

data Config

HTTP/2 configuration.

Constructors

Config

Fields

confWriteBuffer :: Buffer

confBufferSize :: BufferSize

confSendAll :: ByteString -> IO ()

confReadN :: Int -> IO ByteString

confPositionReadMaker :: PositionReadMaker

confTimeoutManager :: Manager

allocSimpleConfig :: Socket -> BufferSize -> IO Config

Making simple configuration whose IO is not efficient. A write buffer is allocated internally.

freeSimpleConfig :: Config -> IO ()

Deallocating the resource of the simple configuration.

HTTP/2 server

type Server = Request -> Aux -> (Response -> [PushPromise] -> IO ()) -> IO ()

Server type. Server takes a HTTP request, should generate a HTTP response and push promises, then should give them to the sending function. The sending function would throw exceptions so that they can be logged.

Request

data Request

Request from client.

Accessing request

requestMethod :: Request -> Maybe Method

Getting the method from a request.

requestPath :: Request -> Maybe Path

Getting the path from a request.

requestAuthority :: Request -> Maybe Authority

Getting the authority from a request.

requestScheme :: Request -> Maybe Scheme

Getting the scheme from a request.

requestHeaders :: Request -> HeaderTable

Getting the headers from a request.

requestBodySize :: Request -> Maybe Int

Getting the body size from a request.

getRequestBodyChunk :: Request -> IO ByteString

Reading a chunk of the request body. An empty ByteString returned when finished.

getRequestTrailers :: Request -> IO (Maybe HeaderTable)

Reading request trailers. This function must be called after getRequestBodyChunk returns an empty.

Aux

data Aux

Additional information.

auxTimeHandle :: Aux -> Handle

Time handle for the worker processing this request and response.

Response

data Response

Response from server.

Creating response

responseNoBody :: Status -> ResponseHeaders -> Response

Creating response without body.

responseFile :: Status -> ResponseHeaders -> FileSpec -> Response

Creating response with file.

responseStreaming :: Status -> ResponseHeaders -> ((Builder -> IO ()) -> IO () -> IO ()) -> Response

Creating response with streaming.

responseBuilder :: Status -> ResponseHeaders -> Builder -> Response

Creating response with builder.

Accessing response

responseBodySize :: Response -> Maybe Int

Getter for response body size. This value is available for file body.

Trailers maker

type TrailersMaker = Maybe ByteString -> IO NextTrailersMaker

Trailers maker. A chunks of the response body is passed with Just. The maker should update internal state with the ByteString and return the next trailers maker. When response body reaches its end, Nothing is passed and the maker should generate trailers. An example:

{-# LANGUAGE BangPatterns #-}
import Data.ByteString (ByteString)
import qualified Data.ByteString.Char8 as C8
import Crypto.Hash (Context, SHA1) -- cryptonite
import qualified Crypto.Hash as CH

-- Strictness is important for Context.
trailersMaker :: Context SHA1 -> Maybe ByteString -> IO NextTrailersMaker
trailersMaker ctx Nothing = return $ Trailers [("X-SHA1", sha1)]
  where
    !sha1 = C8.pack $ show $ CH.hashFinalize ctx
trailersMaker ctx (Just bs) = return $ NextTrailersMaker $ trailersMaker ctx'
  where
    !ctx' = CH.hashUpdate ctx bs

Usage example:

let h2rsp = responseFile ...
    maker = trailersMaker (CH.hashInit :: Context SHA1)
    h2rsp' = setResponseTrailersMaker h2rsp maker

data NextTrailersMaker

Either the next trailers maker or final trailers.

Constructors

NextTrailersMaker TrailersMaker

Trailers [Header]

defaultTrailersMaker :: TrailersMaker

TrailersMake to create no trailers.

setResponseTrailersMaker :: Response -> TrailersMaker -> Response

Setting TrailersMaker to Response.

Push promise

data PushPromise

HTTP/2 push promise or sever push. Pseudo REQUEST headers in push promise is automatically generated. Then, a server push is sent according to promiseResponse.

pushPromise :: ByteString -> Response -> Weight -> PushPromise

Creating push promise. The third argument is traditional, not used.

promiseRequestPath :: PushPromise -> ByteString

Accessor for a URL path in a push promise (a virtual request from a server). E.g. "/style/default.css".

promiseResponse :: PushPromise -> Response

Accessor for response actually pushed from a server.

promiseWeight :: PushPromise -> Weight

Deprecated: Don't use this

Accessor for response weight.

defaultWeight :: Weight

Deprecated: Don't use this

Default weight.

>>> defaultWeight
16

Types

type Path = ByteString

Path.

type Authority = ByteString

For so-called "Host:" header.

type Scheme = ByteString

"http" or "https".

data FileSpec

File specification.

Constructors

FileSpec FilePath FileOffset ByteCount

type FileOffset = Int64

Offset for file.

type ByteCount = Int64

How many bytes to read

RecvN

defaultReadN :: Socket -> IORef (Maybe ByteString) -> Int -> IO ByteString

Naive implementation for readN.

Position read for files

type PositionReadMaker = FilePath -> IO (PositionRead, Sentinel)

Making a position read and its closer.

type PositionRead = FileOffset -> ByteCount -> Buffer -> IO ByteCount

Position read for files.

data Sentinel

Manipulating a file resource.

Constructors

Closer (IO ())
Closing a file resource. Its refresher is automatiaclly generated by the internal timer.

Refresher (IO ())
Refreshing a file resource while reading. Closing the file must be done by its own timer or something.

defaultPositionReadMaker :: PositionReadMaker

Position read based on Handle.